+++ cacti/CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG - cacti
 
-## 1.8.2 / 2021-03-07
+## 1.8.3 / 2021-07-15 / Agent 7.30.0
+
+* [Fixed] Fix example config. See [#9705](https://github.com/DataDog/integrations-core/pull/9705).
+
+## 1.8.2 / 2021-03-07 / Agent 7.27.0
 
 * [Fixed] Bump minimum base package version. See [#8443](https://github.com/DataDog/integrations-core/pull/8443).