+++ rethinkdb/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG - RethinkDB
 
+## 2.0.0 / 2021-08-22 / Agent 7.31.0
+
+* [Removed] Remove documentation specifications. See [#9763](https://github.com/DataDog/integrations-core/pull/9763).
+
+## 1.3.3 / 2021-07-07 / Agent 7.30.0
+
+* [Fixed] Use agent 8 signature. See [#9649](https://github.com/DataDog/integrations-core/pull/9649).
+
 ## 1.3.2 / 2021-01-25 / Agent 7.26.0
 
 * [Fixed] Update logs template with docs feedback. See [#8412](https://github.com/DataDog/integrations-core/pull/8412).